added a lot to the readme and worked on fixing some stuff from the review
