+++ .env.example
@@ -10,17 +10,24 @@ AZURE_CLIENT_ID=
 AZURE_TENANT_ID=
 AZURE_CLIENT_SECRET=
 
+#ASK CJ ABOUT THIS SECTION 
+
 # Get key from https://platform.openai.com/account/api-keys
+# This gives the user access to the OpenAI API, which is necessary for the Agent to function. 
 OPENAI_API_KEY=<YOUR_API_KEY>
 
-# For cloud based testing: https://cloud.griptape.ai
-# For Skatepark testing: http://127.0.0.1:5000 
+# For a structure running in Griptape Cloud, set this variable to: https://cloud.griptape.ai
+# If running against Skatepark, set this environment variable to tell the client which URL Skatepark is listening on.
+# Default is https://127.0.0.1:5000, but is configurable when running gt skatepark start. 
 GT_CLOUD_BASE_URL=
 
-# Get key from https://cloud.griptape.ai/account/api-keys 
+# The Skatepark emulator does not require Griptape Cloud API Key.  
+# Structures in Griptape Cloud require a valid API key to authorize your client to call them. 
+# Generate a key from https://cloud.griptape.ai/account/api-keys 
 GT_CLOUD_API_KEY=<YOUR_API>
 
-#For cloud based testing: get from https://cloud.griptape.ai/knowledge-bases
+#To run the  get from https://cloud.griptape.ai/knowledge-bases
+#For non-cloud based testing: don't know if it's necessary. 
 KNOWLEDGE_BASE_ID=
 
 # Structure ID from the Griptape Cloud or Skatepark
@@ -30,9 +37,20 @@ KNOWLEDGE_BASE_ID=
 # get the id from running 'gt skatepark register --main-file <path-to-main-file>'
 GT_STRUCTURE_ID=
 
-#For Skatepark to simulate a delay in the queue. 
+# When running in Griptape Cloud, the Structure first reports
+# a status of QUEUED before transitioning to RUNNING. Skatepark
+# emulates this behavior by applying a default time to remain 
+# in the QUEUED state in order to allow your client to 
+# handle this state appropriately. By default, Skatepark
+# keeps your Structure QUEUED for 2 seconds.
+# You can override the time that the Structure remains in
+# the QUEUED state within Skatepark by setting this environment
+# variable to the desired time in seconds.
 GT_SKATEPARK_QUEUE_DELAY=2
 
-# If using CDK for conversation memory: get from https://us-east-1.console.aws.amazon.com/lambda/home?region=us-east-1#/functions
+# Allows the user to retrieve the session ID from the CDK in order to interact with the DynamoDB table
+# and use conversation memory. 
+# This is only necessary when running the structure in the Griptape Cloud or in Skatepark. 
+# Get from https://<CDK_DEFAULT_REGION>.console.aws.amazon.com/lambda/home?region=<CDK_DEFAULT_REGION>#/functions
 # Description: Griptape Chatbot Lambda function to invoke a Griptape Structure
 LAMBDA_ENDPOINT=
