--- README_test.md
@@ -1,1 +1,0 @@
-# Griptape Chat Demo
