+++ griptape/chat_cloud/__init__.py
@@ -1,8 +1,6 @@
 from .chat_cloud import Chat_Cloud
-from .chat_cloud import Chat_Local
 
 
 __all__ = [
     "Chat_Cloud",
-    "Chat_Local"
 ]
