+++ griptape/chat_local/__init__.py
@@ -1,0 +1,6 @@
+from .chat_local import Chat_Local
+
+
+__all__ = [
+    "Chat_Local"
+]
